# api-spec/src/main/resources/openapi/stocks.yaml
openapi: 3.0.0F
info:
  title: Stocks Management API
  version: "1.0.0"
  description: API for managing stock investments, fund transfers, and portfolio simulations.
servers:
  - url: http://localhost:8080/api
# - url: http://20.166.52.164/api
#...........................................................
paths:
  /users:
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    get:
      summary: Retrieve all users
      operationId: getAllUsers
      tags:
        - User
      responses:
        '200':
          description: A list of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientModel'
  /users/{id}:
    get:
      summary: Retrieve a user by ID
      operationId: getUserById
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update an existing user
      operationId: updateUser
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: ID of the user to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    delete:
      summary: Delete a user by ID
      operationId: deleteUser
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: User successfully deleted
        '404':
          description: User not found
  /users/by-email/{email}:
    get:
      summary: Find a user by email
      operationId: getUserByEmail
      tags:
        - User
      parameters:
        - name: email
          in: path
          description: The email address of the user to retrieve
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
#...........................................................
...
components:
  schemas:
...
    ClientModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          description: The user’s app nickname
        email:
          type: string
          format: email
          description: The user’s email address
        role:
          type: string
          enum:
            - USER
            - ADMIN
          description: The user’s role in the system
        iban:
          type: string
          description: The user’s bank account IBAN
        bankAccountBalance:
          type: number
          format: double
          description: The current balance of the user’s bank account
        investmentAccountBalance:
          type: number
          format: double
          description: The current balance of the user’s investment account
      required:
        - username
        - email
        - role
        - iban
        - bankAccountBalance
        - investmentAccountBalance
#...........................................................
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        investmentAccount:
          $ref: '#/components/schemas/InvestmentAccount'
        bankAccount:
          $ref: '#/components/schemas/BankAccount'
#...........................................................
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - role
      properties:
        username:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        bankAccountBalance:
          type: number
#...........................................................
    UserRole:
      type: string
      enum:
        - USER
        - ADMIN
#...........................................................
    BankAccount:
      type: object
      properties:
        iban:
          type: string
        balance:
          type: number
          format: double
#...........................................................
    InvestmentAccount:
      type: object
      properties:
        id:
          type: integer
          format: int64
        balance:
          type: number
          format: double
        portfolio:
          $ref: '#/components/schemas/Portfolio'
#...........................................................
    Portfolio:
      type: object
      properties:
        id:
          type: integer
          format: int64
        holdings:
          type: array
          items:
            $ref: '#/components/schemas/StockHolding'
#...........................................................
    StockHolding:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sharesOwned:
          type: number
          format: double
        stock:
          $ref: '#/components/schemas/Stock'
#...........................................................
    Stock:
      type: object
      properties:
        id:
          type: integer
          format: int64
        symbol:
          type: string
        currency:
          type: string
        currentPrice:
          type: number
          format: double
        volatility:
          type: number
          format: double
        dividendYield:
          type: number
          format: double
        expectedReturn:
          type: number
          format: double
#...........................................................
...